add new project and move some base-layer stuff over
