+++ src/Cli/Microsoft.DotNet.Cli.CommandLine/DynamicArgument.cs
@@ -1,0 +1,8 @@
+﻿using System.CommandLine;
+using System.CommandLine.StaticCompletions;
+
+namespace Microsoft.DotNet.Cli.CommandLine;
+
+public class DynamicArgument<T>(string name) : Argument<T>(name), IDynamicArgument
+{
+}
