+++ src/Cli/dotnet/Commands/Pack/PackCommandParser.cs
@@ -2,9 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
+using Microsoft.DotNet.Cli.CommandLine;
 using Microsoft.DotNet.Cli.Commands.Build;
 using Microsoft.DotNet.Cli.Commands.Restore;
-using Microsoft.DotNet.Cli.Extensions;
 using NuGet.Versioning;
 
 namespace Microsoft.DotNet.Cli.Commands.Pack;
