+++ src/Cli/dotnet/Commands/Package/List/PackageListCommand.cs
@@ -6,6 +6,7 @@
 using System.CommandLine;
 using Microsoft.DotNet.Cli.Commands.Hidden.List;
 using Microsoft.DotNet.Cli.Commands.NuGet;
+using Microsoft.DotNet.Cli.CommandLine;
 using Microsoft.DotNet.Cli.Extensions;
 using Microsoft.DotNet.Cli.Utils;
 using System.Globalization;
