+++ src/Cli/dotnet/Commands/Package/List/PackageListCommandParser.cs
@@ -4,7 +4,7 @@
 #nullable disable
 
 using System.CommandLine;
-using Microsoft.DotNet.Cli.Extensions;
+using Microsoft.DotNet.Cli.CommandLine;
 
 namespace Microsoft.DotNet.Cli.Commands.Package.List;
