+++ src/Cli/dotnet/Commands/Package/Remove/PackageRemoveCommand.cs
@@ -5,7 +5,7 @@ using System.CommandLine;
 using System.Diagnostics;
 using Microsoft.DotNet.Cli.Commands.NuGet;
 using Microsoft.DotNet.Cli.Commands.Run;
-using Microsoft.DotNet.Cli.Extensions;
+using Microsoft.DotNet.Cli.CommandLine;
 using Microsoft.DotNet.Cli.Utils;
 
 namespace Microsoft.DotNet.Cli.Commands.Package.Remove;
